requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pytest>=7.4.0
urllib3>=2.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
